split BIEST into header and implementation; split off ParaVIEW interface into its own paraview.hpp
