+++ include/biest/boundary_integ_op.hpp
@@ -2,6 +2,7 @@
 #define _BOUNDARY_INTEG_OP_HPP_
 
 #include <biest/surface.hpp>
+#include <biest/surface_op.hpp>
 #include <sctl.hpp>
 
 namespace biest {
